+++ History.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Fix handling of `undefined` in `res.jsonp`
   * Fix incorrect middleware execution with unanchored `RegExp`s
   * Fix `res.jsonp(obj, status)` deprecation message
   * Fix typo in `res.is` JSDoc